Removing uses of SocketCreator from MemberIdentifierImpl
Moving readEssentialData to InternalDistributedMember, because it relies on
SocketCreator, but it is only used for this feature of
InternalDistributedMember.

Changing the places that used SocketCreator.getHostName in MemberIdentifierImpl
to not try to modify the hostname during deserialization. Whatever the hostName
field was on the remote side is what it will be on the local side.
